rename fb_state to state
